HBX-1607: Move class 'org.hibernate.cfg.binder.PropertyBinder' to 'org.hibernate.tool.internal.reveng.PropertyBinder'
